requirements.txt
fastapi
uvicorn
google-cloud-firestore
firebase-admin
pyserial
python-dotenv